refactor: move logger to common-utils lib
