+++ apps/cowswap-frontend/src/legacy/hooks/usePriceImpact/logger.ts
@@ -1,3 +1,3 @@
-import { createCowLogger } from 'common/utils/logger'
+import { createCowLogger } from '@cowprotocol/common-utils'
 
 export const logPriceImpact = createCowLogger('PriceImpact')
