+++ apps/cowswap-frontend/src/modules/captcha/logger.ts
@@ -1,3 +1,3 @@
-import { createCowLogger } from 'common/utils/logger'
+import { createCowLogger } from '@cowprotocol/common-utils'
 
 export const logCaptcha = createCowLogger('Captcha')
