+++ libs/core/package.json
@@ -27,6 +27,7 @@
     "@cowprotocol/cow-sdk": "9.2.2",
     "@cowprotocol/cms": "https://registry.npmjs.org/@cowprotocol/cms/-/cms-0.11.0.tgz",
     "@cowprotocol/common-const": "workspace:*",
+    "@cowprotocol/common-utils": "workspace:*",
     "@safe-global/api-kit": "4.0.1",
     "@safe-global/protocol-kit": "1.3.0",
     "@safe-global/types-kit": "3.0.0",
